* Broke off audio init code into its own file...to preserve my sanity
