+++ src/engine/audio/CMakeLists.txt
@@ -1,9 +1,11 @@
 
 list(APPEND GAME_SOURCES
+	"${CMAKE_CURRENT_SOURCE_DIR}/init.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/sound.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/sound_game.cpp"
 )
 list(APPEND EDITOR_SOURCES
+	"${CMAKE_CURRENT_SOURCE_DIR}/init.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/sound.cpp"
 )
